new_requirement.txt
alabaster==0.7.12
appdirs==1.4.4
argon2-cffi==20.1.0
arrow==0.17.0
asgiref==3.3.0
astroid==2.4.2
attrs==20.3.0
Babel==2.8.0
backcall==0.2.0
binaryornot==0.4.4
black==20.8b1
certifi==2020.11.8
cffi==1.14.3
cfgv==3.2.0
chardet==3.0.4
click==7.1.2
cookiecutter==1.7.2
coreapi==2.3.3
coreschema==0.0.4
coverage==5.3
cryptography==3.2.1
dataclasses==0.7
decorator==4.4.2
defusedxml==0.6.0
distlib==0.3.1
Django==3.0.11
django-allauth==0.43.0
django-cors-headers==3.5.0
django-coverage-plugin==1.8.0
django-crispy-forms==1.9.2
django-debug-toolbar==3.1.1
django-environ==0.4.5
django-extensions==3.0.9
django-model-utils==4.0.0
django-redis==4.12.1
django-stubs==1.7.0
djangorestframework==3.12.2
docutils==0.16
factory-boy==3.0.1
Faker==4.14.2
filelock==3.0.12
flake8==3.8.4
flake8-isort==4.0.0
h11==0.11.0
hiredis==1.1.0
identify==1.5.9
idna==2.10
imagesize==1.2.0
importlib-metadata==2.0.0
importlib-resources==3.3.0
iniconfig==1.1.1
ipdb==0.13.4
ipython==7.16.1
ipython-genutils==0.2.0
isort==5.6.4
itypes==1.2.0
jedi==0.17.2
Jinja2==2.11.2
jinja2-time==0.2.0
lazy-object-proxy==1.4.3
livereload==2.6.3
MarkupSafe==1.1.1
mccabe==0.6.1
mypy==0.790
mypy-extensions==0.4.3
nodeenv==1.5.0
oauthlib==3.1.0
packaging==20.4
parso==0.7.1
pathspec==0.8.1
pexpect==4.8.0
pickleshare==0.7.5
Pillow==8.0.1
pluggy==0.13.1
poyo==0.5.0
pre-commit==2.8.2
prompt-toolkit==3.0.8
psycopg2-binary==2.8.6
ptyprocess==0.6.0
py==1.9.0
pycodestyle==2.6.0
pycparser==2.20
pyflakes==2.2.0
Pygments==2.7.2
PyJWT==1.7.1
pylint==2.6.0
pylint-django==2.3.0
pylint-plugin-utils==0.6
pyparsing==2.4.7
pytest==6.1.2
pytest-django==4.1.0
pytest-sugar==0.9.4
python-dateutil==2.8.1
python-slugify==4.0.1
python3-openid==3.2.0
pytz==2020.4
PyYAML==5.3.1
redis==3.5.3
regex==2020.10.28
requests==2.24.0
requests-oauthlib==1.3.0
six==1.15.0
snowballstemmer==2.0.0
Sphinx==3.3.0
sphinx-autobuild==2020.9.1
sphinxcontrib-applehelp==1.0.2
sphinxcontrib-devhelp==1.0.2
sphinxcontrib-htmlhelp==1.0.3
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.3
sphinxcontrib-serializinghtml==1.1.4
sqlparse==0.4.1
termcolor==1.1.0
testfixtures==6.15.0
text-unidecode==1.3
toml==0.10.2
tornado==6.1
traitlets==4.3.3
typed-ast==1.4.1
typing-extensions==3.7.4.3
uritemplate==3.0.1
urllib3==1.25.11
uvicorn==0.12.2
virtualenv==20.1.0
watchgod==0.6
wcwidth==0.2.5
Werkzeug==1.0.1
wrapt==1.12.1
wsproto==0.15.0
zipp==3.4.0
